Start organising/adding table-generation functions
